refactor(chat): split structured cards out of message-content
Extract AutomationProposalCard, ConnectionRequiredCard, and QuickReplies
from message-content.tsx into their own files so each component has a
single-responsibility home and message-content.tsx stays focused on
MessageContentWithAuth (~95 LOC).

Files added:
  - automation-proposal-card.tsx (~50 LOC)
  - connection-required-card.tsx (~95 LOC)
  - quick-replies.tsx (~30 LOC)

Updated ai-chat-box.tsx to import QuickReplies from its new location.
