+++ packages/web/src/app/routes/chat-with-ai/components/quick-replies.tsx
@@ -1,0 +1,29 @@
+import { motion } from 'motion/react';
+
+export function QuickReplies({
+  replies,
+  onSend,
+}: {
+  replies: string[];
+  onSend: (text: string, files?: File[]) => void;
+}) {
+  if (replies.length === 0) return null;
+
+  return (
+    <div className="flex flex-wrap gap-2 py-2">
+      {replies.map((reply, i) => (
+        <motion.button
+          key={reply}
+          type="button"
+          onClick={() => onSend(reply)}
+          className="px-3 py-1.5 text-sm rounded-full border bg-background hover:bg-muted transition-colors cursor-pointer"
+          initial={{ opacity: 0, y: 6 }}
+          animate={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.25, delay: i * 0.06 }}
+        >
+          {reply}
+        </motion.button>
+      ))}
+    </div>
+  );
+}
